Print errors, handle strings starting with floats
Print the errors with the record that caused the error so I can see what
records I need to fix or de-duplicate.

Remove a lot of the debug clutter.

The float state in budget_csv:record/4 didn't check if the current field
turned from a float into a string: e.g. 1.23abc; once the "a" is
encountered the float state should toss this field to the string state.