Fix CSV date, print counts, add SQL
The first date field was being compared as a string to see if it was
greater than twelve (i.e. a year column instead of a month column). It
needs to be an integer.

Print out the number of records, how many were inserted and how many had
errors.

Add SQL scripts to drop and create the database. I'm intending to allow
each transaction to have mutliple categories. There is a rules table to
match against descriptions to automatically categorize transactions.