Get JSONP working
Getting some in-page data loading going.

The root path is handled by the budget_web_h handler (see
budget_app.erl) for the routing. The budget web handler expects three
querystring values: callback, from and to; the handler fetches
transactions between from and to (which are dates), converts them to
JSON and wraps them in a JavaScript function with the name specified in
the callback querystring value.

The web page has a starter script with a q/2 function that will replace
any existing "fetch" script with a new one fetched from the web server.
Each new fetch script should have the callback specified in the source
address of the new fetch script element.