Move category datalist input to table column
My idea is to have the list of categories at the end of each record in
the table so the user can select new categories for each transaction. As
soon as the user selects a complete category name then that category is
added to the transaction and the database is updated. I'm planning on
having some kind of "delete" icon or element on each category (e.g. a
mouse-hover trash can icon or 'x') so the user can easily add or delete
categories.