Add category "+" button which looks up category ID
This is work towards dynamically adding categories to transactions. I
was going to add a category as soon as the text input with datalist was
set to a valid cateogry; however, if you're typing in a category it
would be weird to be updating the auto-complete matches until you struck
a value category and then all of the sudden it's added. Click on a
category would be fine and I could distinguish the two types of input
events but the if you type the complete category you have to then click
on it, which would be weird. A "+" button is pretty universal.