Add category to transaction when "+" is clicked
Added an index to the transaction_category table so that transactions
don't get duplicate categories.

I'm using a regular XMLHttpRequest to get the upated categories for a
transaction but I'm not sure if I shouldn't just use JSONP like I do to
fetch the categories in the first place. JSONP is certainly simpler.