Fix POST for new Tx-Cat: can only read body once
I was reading the body in resource_exists/2 and then trying to re-read
it again later. Instead I just stuck it in the State.

FYI, the "State" is set by budget_app.erl in the routes. Currently I'm
setting all handler initial states to [].

Now that I've filled in resource_exists/2 I'm getting a 404 not found
when I go to fetch all transaction_category records for a transaction
that currently doesn't have any categories. I'm thinking I'll pass a
404-specific callback to the XHR gear for that case.