Fix POST to transaction_category
Fixed resource_exists logic. Instead of looking for existing
transaction_category on GET just look to see if the transaction exists.
I don't want to throw an error because the list of tx categories is
empty; but I do want to throw an error if the tx cats list is requested
for a tx that doesn't exist.

Also fix the DB value comparisons. I haven't completely worked out when
I'm getting binary and when I'm getting integer values.