WIP: POST to copy tx, GET to fetch it back
Working on using a POST with the parent ID to create a child copy of it
for splitting transactions. Still working through the bugs getting the
POST to work. Once POST works I need to fetch the child back with a GET
and then create a new row for it in the table.

The first split will create two child transactions then the button will
switch to "split again" and will create one more transaction at a time.

The order of the query to fetch transactions has to order them so that
the parent comes first followed by the children in child_number order.

Also moved the non-header table rows to a table body instead of the
table head.

I still need to clear out the amounts for the child records. The user
will have to fill them in. That means I'll need to changed the amounts
to text inputs when the user clicks on them.