Add new child rows to table upon split
Fixed up the POST callback so that a unique script will be created for
each of the two JSONP results that will come back from the web server. I
had to add the source to the scripts manually unless I wanted to change
the transaction POST handler to _not_ return {true, URI} but return the
URI instead, which I could then call.