Delete parent categories on split, add app modules
When a transaction is split its categories are deleted since the child
transactions will be categorized individually. I could leave the
categories on the parent transaction and just not include them in any
queries or calculations but that would probably just be confusing.

Add the web handler modules to the modules key of the .app.src file so i
don't have to load them manually. I spent so much time manually
compiling the handlers while the node was running that I couldn't figure
out why the handlers couldn't be found on a freshly started server when
I tried to load the transactions on the web page.