Unparent a former parent with no children
If a tx is split it becomes a parent to 1 or more tx children. It starts
with 2 but you can delete 1 at a time so it could have one. I should
probably fix that; I can't think of a reaon why the user would ever need
to go down to 1 child row: if another child had the wrong category or
amount it could just be changed. Anyway, if you delete all of the
children from a parent then it's no longer a parent and needs to get
it's controls back for adding categories. Also, the "split again" button
needs to go back to "split" and the parent colour needs to be removed.