Put tx.parent field back in to control child props
I updated the transaction fetch query from t.* to a list of specific
fields but I left out the parent field which isn't shown in the table.
The parent field is the ID of the parent which implies that the record
is a child. In JavaScript any string value that is not undefined or ""
has a truth-value of True. We can use this truth value in:
 "let is_child = obj.parent" - if parent is undefined or "" then
is_child is false.