Remove category and delete controls on split
Parent transactions shouldn't have categories or be deleted.

There's a bug here: if you split a transaction then delete the children,
then split it again, the category controls don't disappear. Then, if you
delete the children again an extra set of category controls gets added.
So, something is goofy in the "unparent" gear.